Finished TestItemService (passing all tests)

+++ GroceryStore-Backend/src/main/java/ca/mcgill/ecse321/GroceryStore/service/ItemService.java
@@ -19,17 +19,24 @@ public class ItemService {
     public Item createItem(String name, boolean purchasable, int price, String description, int stock) {
         ArrayList<String> errorMessages = new ArrayList<>();
 
-        if (name == null || name.trim().length() == 0) errorMessages.add("Name can't be empty");
+        if (name == null || name.trim().length() == 0) errorMessages.add("Name can't be empty.");
 
         if (description == null || description.trim().length() == 0) errorMessages.add("Description can't be empty.");
+
+        if (price == 0) errorMessages.add("Item's price can't be 0.");
+
+        if (stock == 0) errorMessages.add("Item's stock can't be 0.");
+
+        if (price < 0) errorMessages.add("Item's price can't be negative.");
+
+        if (stock < 0) errorMessages.add("Item's stock can't be negative.");
 
         if (errorMessages.isEmpty()) {
             if (itemRepository.existsById(name)) {
                 errorMessages.add("An identical Item already exists.");
             }
         }
-
-
+        
         if (errorMessages.size() > 0) throw new IllegalArgumentException(String.join(" ", errorMessages));
 
         Item item = new Item();
@@ -45,6 +52,7 @@ public class ItemService {
 
     @Transactional
     public Item getItem(String name) {
+        if (name == null || name.trim().length() == 0) throw new IllegalArgumentException("A name parameter is needed.");
         Item item = itemRepository.findByName(name);
         if (item == null) throw new IllegalArgumentException("The Item with name: " + name + " was not found in the database.");
         return item;
@@ -57,6 +65,7 @@ public class ItemService {
 
     @Transactional
     public void deleteItem(String name) {
+        if (name == null || name.trim().length() == 0) throw new IllegalArgumentException("A name parameter is needed.");
         Item item = itemRepository.findByName(name);
         if (item == null) throw new IllegalArgumentException("The item with name " + name + " does not exist.");
         else itemRepository.deleteById(name);
